+++ src/matrices_operations/matrix_add.hh
@@ -10,6 +10,6 @@
 #include <cmath>
 #include <cuda_runtime.h>
 
-float* mat_add(float* A, float* B, int NA, int MA, int NB, int MB);
+double* mat_add(double* A, double* B, int NA, int MA, int NB, int MB);
 
 #endif //GPGPU_MATRIX_ADD_HH
