+++ src/matrices_operations/matrix_mult.cu
@@ -7,7 +7,7 @@
 
 
 
-float* mat_mult(float* A, float* B, int NA, int MA, int NB, int MB)
+double* mat_mult(double* A, double* B, int NA, int MA, int NB, int MB)
 {
     if (MA != NB) {
         std::cerr << "shape error" << std::endl;
@@ -21,29 +21,29 @@ float* mat_mult(float* A, float* B, int NA, int MA, int NB, int MB)
 
 
     // Allocate memory on the device
-    float* d_A;
-    float* d_B;
-    float* d_C;
-    auto* C = (float*)malloc(SIZE_C * sizeof(float));
+    double* d_A;
+    double* d_B;
+    double* d_C;
+    auto* C = (double*)malloc(SIZE_C * sizeof(double));
 
-    cudaMalloc(&d_A, SIZE_A * sizeof(float));
-    cudaMalloc(&d_B, SIZE_B * sizeof(float));
-    cudaMalloc(&d_C, SIZE_C * sizeof(float));
+    cudaMalloc(&d_A, SIZE_A * sizeof(double));
+    cudaMalloc(&d_B, SIZE_B * sizeof(double));
+    cudaMalloc(&d_C, SIZE_C * sizeof(double));
 
 
     // Copy to device
-    rc = cudaMemcpy(d_A, &A[0], SIZE_A * sizeof(float), cudaMemcpyHostToDevice);
+    rc = cudaMemcpy(d_A, &A[0], SIZE_A * sizeof(double), cudaMemcpyHostToDevice);
     if (rc)
         std::cout << "error memcpy\n";
-    cudaMemcpy(d_B, &B[0], SIZE_B * sizeof(float), cudaMemcpyHostToDevice);
-    cudaMemset(d_C, 0, SIZE_C * sizeof(float));
+    cudaMemcpy(d_B, &B[0], SIZE_B * sizeof(double), cudaMemcpyHostToDevice);
+    cudaMemset(d_C, 0, SIZE_C * sizeof(double));
 
     // call the kernel
     matrixMultiplication(d_A, d_B, d_C, NA, MA, NB, MB);
     cudaDeviceSynchronize();
 
     // copy memory back to host
-    cudaMemcpy(&C[0], d_C, SIZE_C * sizeof(float), cudaMemcpyDeviceToHost);
+    cudaMemcpy(&C[0], d_C, SIZE_C * sizeof(double), cudaMemcpyDeviceToHost);
 
     return C;
 }
